Bug 1277581 – Incidental move of Homepage Accessors into Accessors group.
